docs(frontend-docs): clean up story controls
