+++ apps/frontend-docs/design-system-vue/src/stories/Link.stories.ts
@@ -1,7 +1,5 @@
 import {
-  copyFrontendDocsSnippetToClipboard,
   createFrontendDocsComponentSnippet,
-  createFrontendDocsHighlightedSnippetHtml,
   createFrontendDocsPlaygroundParameters,
   createVueStoryPreview,
   createVueStorySourcePanel,
@@ -20,7 +18,8 @@ import {
 } from '@hoite-dev/ui';
 import { Link } from '@hoite-dev/ui-vue';
 import type { ArgTypes, Meta, StoryObj } from '@storybook/vue3-vite';
-import { computed, defineComponent, type PropType, ref } from 'vue';
+import { computed, defineComponent, type PropType } from 'vue';
+import { createVueSnippetCopyState } from './vueSnippetCopyState';
 
 const linkPlaygroundControlNames = ['children', 'href', 'appearance', 'target', 'rel'] as const;
 
@@ -169,16 +168,7 @@ const LinkPlaygroundPreview = defineComponent({
         ],
       }),
     );
-    const copyButtonLabel = ref('Copy code');
-    const copySnippet = async () => {
-      copyButtonLabel.value = 'Copying';
-      copyButtonLabel.value = (await copyFrontendDocsSnippetToClipboard(snippet.value))
-        ? 'Copied'
-        : 'Copy error';
-    };
-    const highlightedSnippet = computed(() =>
-      createFrontendDocsHighlightedSnippetHtml(snippet.value),
-    );
+    const { copyButtonLabel, copySnippet, highlightedSnippet } = createVueSnippetCopyState(snippet);
 
     return {
       copyButtonLabel,
