+++ apps/frontend-docs/design-system-vue/src/stories/Typography.stories.ts
@@ -1,7 +1,5 @@
 import {
-  copyFrontendDocsSnippetToClipboard,
   createFrontendDocsComponentSnippet,
-  createFrontendDocsHighlightedSnippetHtml,
   createFrontendDocsPlaygroundParameters,
   createVueStoryPreview,
   createVueStorySourcePanel,
@@ -17,7 +15,8 @@ import {
 } from '@hoite-dev/ui';
 import { Typography } from '@hoite-dev/ui-vue';
 import type { ArgTypes, Meta, StoryObj } from '@storybook/vue3-vite';
-import { computed, ref } from 'vue';
+import { computed } from 'vue';
+import { createVueSnippetCopyState } from './vueSnippetCopyState';
 
 const defaultTagOption = 'Default variant tag';
 const variantKeys = Object.keys(typographyVariantConfig) as TypographyVariant[];
@@ -113,16 +112,8 @@ export const Playground: Story = {
           ],
         }),
       );
-      const copyButtonLabel = ref('Copy code');
-      const copySnippet = async () => {
-        copyButtonLabel.value = 'Copying';
-        copyButtonLabel.value = (await copyFrontendDocsSnippetToClipboard(snippet.value))
-          ? 'Copied'
-          : 'Copy error';
-      };
-      const highlightedSnippet = computed(() =>
-        createFrontendDocsHighlightedSnippetHtml(snippet.value),
-      );
+      const { copyButtonLabel, copySnippet, highlightedSnippet } =
+        createVueSnippetCopyState(snippet);
 
       return {
         args,
